requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SocketIO==5.5.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
opencv-python==4.11.0.86
pyserial==3.5
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
